Network form & API fixes and improvements (#2445)
Added additional field to IP address: `dhcp_expose`. When it's checked, trio (hostname, mac, ip) will be exposed in DHCP config. There are some constraints for this (when `dhcp_expose` is selected) in GUI and API:
* mac, hostname or ip could not be changed
* `dhcp_expose` could be changed only by transition

So it's easy to add entry to DHCP, but more complicated to remove it from DHCP (requires executing transition).