Fix commiting async job to worker (#2703)
Fix commiting async job to worker

When async transition (job) is scheduled to worker, transaction has to be already commited to allow worker to fetch job from database. This cannot be done in atomic block (especially when whole request is atomic), so from now transition view is non-atomic (gui + api) for async transitions and is commited manually after creating job object in database.