Event processing for change management (#2998)
* Make operations identifiable by ticket id

Change management workflows that involve usage of an external ticket
systems allow to identify every single operation by its external ID.
In order to do that a single operation should be ralated to a single
ticket.

This patch sets a unique property for an operation's ticket_id
attribute so identifying tickets by their external identifiers is
possible.

* Extend the list of available operation statuses

This patch adds the following statuses:

 - reopened
 - todo
 - blocked

* Event processing for change management

This patch adds a generic subscriber for receiving and recording
change management events and a Jira processor for it. Base object
loading logic should be implemented for every secific case so it
is not included.