Refactor operation status
List of possible operation statuses along with their meaning is
a flexible matter, especially in cases, when an external change
management software is used. Given that, having to rewrite and
release a new version of Ralph in order to adapt to the changes
is not a way to go.

This patch removes a hard-coded list of statuses from Ralph and
replaces it with a reference to a manageable entity called
OperationStatus. Those entities are manageable botj over the API
and the UI.

The change management listener's behavior was also set to create
new operation types in responce to incomming messages instead of
ignoring operations that have an unknown status.