Introduce cloud syncronisation
This patch extends cloud providers to contain necessary settings
for cloud syncronisation such as switchable option for enabling
or disabling the syncronisation for every specific provider and
pluggable syncronisation event processor.

The cloud syncronisation endpoint accepts POST requests for every
cloud provider for which the syncronisation is enabled and an event
processor is selected. Events are routed for necessary processors
in the context of a selected cloud provider.